treewide: fix even more warnings[3/N]

+++ arch/arm/mach-msm/include/mach/iommu_hw-v1.h
@@ -58,7 +58,7 @@
 #define SET_FIELD(addr, mask, shift, v) \
 do { \
 	u64 t = readl_relaxed(addr); \
-	writel_relaxed((t & ~((mask) << (shift))) + (((v) & \
+	writel_relaxed((t & ~((u64)(mask) << (shift))) + (((v) & \
 			(mask)) << (shift)), addr); \
 } while (0)
